Split off duplicates/unique tests into separate file
